+++ tests/unit/test_models.py
@@ -6,7 +6,8 @@ import pybamm
 import pytest
 
 import pybop
-from examples.standalone.model import ExponentialDecay
+from examples.standalone.model import ExponentialDecay as StandaloneDecay
+from pybop.models.lithium_ion.basic_SPMe import convert_physical_to_grouped_parameters
 
 
 class TestModels:
@@ -27,23 +28,28 @@ class TestModels:
                 {"number of MSMR reactions": ("6", "4")},
             ),
             (pybop.lithium_ion.WeppnerHuggins, "Weppner & Huggins model", None),
+            (
+                pybop.lithium_ion.GroupedSPMe,
+                "Grouped Single Particle Model with Electrolyte",
+                None,
+            ),
+            (
+                pybop.lithium_ion.GroupedSPMe,
+                "Grouped Single Particle Model with Electrolyte",
+                {"surface form": "differential"},
+            ),
             (pybop.empirical.Thevenin, "Equivalent Circuit Thevenin Model", None),
         ],
     )
     @pytest.mark.unit
     def test_model_classes(self, model_class, expected_name, options):
-        model = model_class(options=options)
+        parameter_set = pybop.ParameterSet(
+            params_dict={"Nominal cell capacity [A.h]": 5.12}
+        )
+        model = model_class(options=options, parameter_set=parameter_set)
         assert model.pybamm_model is not None
         assert model.name == expected_name
-
-        # Test initialisation with kwargs
-        if model_class is pybop.lithium_ion.MSMR:
-            # Reset the options to cope with a bug in PyBaMM v23.9 msmr.py:23 which is fixed in v24.1
-            options = {"number of MSMR reactions": ("6", "4")}
-        parameter_set = pybop.ParameterSet(
-            params_dict={"Nominal cell capacity [A.h]": 5}
-        )
-        model = model_class(options=options, build=True, parameter_set=parameter_set)
+        assert model.parameter_set["Nominal cell capacity [A.h]"] == 5.12
 
     @pytest.fixture(
         params=[
@@ -53,6 +59,8 @@ class TestModels:
             pybop.lithium_ion.MPM(),
             pybop.lithium_ion.MSMR(options={"number of MSMR reactions": ("6", "4")}),
             pybop.lithium_ion.WeppnerHuggins(),
+            pybop.lithium_ion.GroupedSPMe(),
+            pybop.lithium_ion.GroupedSPMe(options={"surface form": "differential"}),
             pybop.empirical.Thevenin(),
         ]
     )
@@ -83,7 +91,9 @@ class TestModels:
             model.predict(None, None)
 
         # Test new_copy() without pybamm_model
-        if not isinstance(model, pybop.lithium_ion.MSMR):
+        if not isinstance(
+            model, (pybop.lithium_ion.MSMR, pybop.lithium_ion.GroupedSPMe)
+        ):
             new_model = model.new_copy()
             assert new_model.pybamm_model is not None
             assert new_model.parameter_set is not None
@@ -92,7 +102,12 @@ class TestModels:
     def test_predict_with_inputs(self, model):
         # Define inputs
         t_eval = np.linspace(0, 10, 100)
-        if isinstance(model, (pybop.lithium_ion.EChemBaseModel)):
+        if isinstance(model, (pybop.lithium_ion.GroupedSPMe)):
+            inputs = {
+                "Negative electrode relative porosity": 0.52,
+                "Positive electrode relative porosity": 0.63,
+            }
+        elif isinstance(model, (pybop.lithium_ion.EChemBaseModel)):
             if model.pybamm_model.options["working electrode"] == "positive":
                 inputs = {
                     "Positive electrode active material volume fraction": 0.63,
@@ -272,11 +287,14 @@ class TestModels:
                 atol=1e-5,
             )
 
+    @pytest.mark.parametrize(
+        "model_cls", [StandaloneDecay, pybop.ExponentialDecayModel]
+    )
     @pytest.mark.unit
-    def test_reinit(self):
+    def test_reinit(self, model_cls):
         k = 0.1
         y0 = 1
-        model = ExponentialDecay(pybamm.ParameterValues({"k": k, "y0": y0}))
+        model = model_cls(pybamm.ParameterValues({"k": k, "y0": y0}))
 
         with pytest.raises(
             ValueError, match="Model must be built before calling get_state"
@@ -291,17 +309,20 @@ class TestModels:
         state = model.reinit(inputs=[1])
         np.testing.assert_array_almost_equal(state.as_ndarray(), np.array([[y0]]))
 
-        model = ExponentialDecay(pybamm.ParameterValues({"k": k, "y0": y0}))
+        model = model_cls(pybamm.ParameterValues({"k": k, "y0": y0}))
         with pytest.raises(
             ValueError, match="Model must be built before calling reinit"
         ):
             model.reinit(inputs={})
 
+    @pytest.mark.parametrize(
+        "model_cls", [StandaloneDecay, pybop.ExponentialDecayModel]
+    )
     @pytest.mark.unit
-    def test_simulate(self):
+    def test_simulate(self, model_cls):
         k = 0.1
         y0 = 1
-        model = ExponentialDecay(pybamm.ParameterValues({"k": k, "y0": y0}))
+        model = model_cls(pybamm.ParameterValues({"k": k, "y0": y0}))
         model.build()
         model.signal = ["y_0"]
         inputs = {}
@@ -311,7 +332,7 @@ class TestModels:
         np.testing.assert_array_almost_equal(solved["y_0"].data, expected, decimal=5)
 
         with pytest.raises(ValueError):
-            ExponentialDecay(n_states=-1)
+            model_cls(n_states=-1)
 
     @pytest.mark.unit
     def test_simulateEIS(self):
@@ -568,3 +589,33 @@ class TestModels:
             dataset_2["Current function [A]"].data,
             atol=1e-8,
         )
+
+    @pytest.mark.unit
+    def test_grouped_SPMe(self):
+        with pytest.warns(UserWarning) as record:
+            model = pybop.lithium_ion.GroupedSPMe(
+                unused_kwarg=0, options={"unused option": 0}
+            )
+            assert "The input model_kwargs" in str(record[0].message)
+            assert "are not currently used by the GroupedSPMe." in str(
+                record[0].message
+            )
+
+        parameter_set = pybop.ParameterSet.pybamm("Chen2020")
+        parameter_set["Electrolyte diffusivity [m2.s-1]"] = 1.769e-10
+        parameter_set["Electrolyte conductivity [S.m-1]"] = 0.9487
+
+        grouped_parameter_set = convert_physical_to_grouped_parameters(parameter_set)
+        model = pybop.lithium_ion.GroupedSPMe(parameter_set=grouped_parameter_set)
+
+        with pytest.raises(
+            ValueError, match="GroupedSPMe can currently only accept an initial SoC."
+        ):
+            model.set_initial_state({"Initial open-circuit voltage [V]": 3.7})
+
+        model.set_initial_state({"Initial SoC": 1.0})
+        res = model.predict(t_eval=np.linspace(0, 10, 100))
+        assert len(res["Voltage [V]"].data) == 100
+
+        variable_list = model.pybamm_model.default_quick_plot_variables
+        assert isinstance(variable_list, list)